Fix enabling testnet issue
